review: Copilotレビューコメント対応 (createConverter切り出し・playwright修正)
- createConverter を image_upload.js から createConverter.js に切り出し、
  lib/distribute.js がサーバーモジュールの副作用なしで参照できるようにする
- lib/distribute.js でループ外で一度だけ createConverter を解決するよう修正
- playwright.config.ts の webServer.command を pnpm 依存から node bin/server に変更
- docs/dictionary.md と test-tree.md の src 参照先を新モジュールに更新
