+++ lib/distribute.js
@@ -9,7 +9,7 @@ import config from './config.js'
 import { cacheDir } from './dir.js'
 import { applyTemplate } from './applyTemplate.js'
 import { distributeImages } from './imageDistributor.js'
-import { createConverter } from '../packages/editor/server/image_upload.js'
+import { createConverter } from '../packages/editor/server/createConverter.js'
 
 const indexFile = `${cacheDir}/index.json`
 
@@ -44,12 +44,12 @@ const distribute = async (data, srcDir, distDir) => {
     })
   }
   const distributeRaw = config.distribute_raw.split(',')
+  const imageConverter = config.image_converter ? await createConverter(config.image_converter) : null
   for (const copyDir of distributeRaw) {
     if (!existsSync(`${cacheDir}/${copyDir}`)) {
       continue
     }
-    if (config.image_converter) {
-      const imageConverter = await createConverter(config.image_converter)
+    if (imageConverter) {
       await distributeImages(`${cacheDir}/${copyDir}/`, `${distDir}/${copyDir}/`, imageConverter)
       console.log(styleText('green', '[convert]'), `${cacheDir}/${copyDir}/ => ${distDir}/${copyDir}/`)
     } else {
